requirements.txt
streamlit==1.38.0
requests==2.32.3
pandas==2.2.2
fpdf==1.7.2
python-docx==1.1.2
Pillow==10.4.0
beautifulsoup4==4.12.3
lxml==5.3.0
XlsxWriter==3.2.0